+++ GUI/src/i18n/et/common.json
@@ -2,6 +2,17 @@
   "global": {
     "title": "Analyticsi moodul"
   },
+  "general": {
+    "period": "Periood",
+    "metric": "Mõõdik",
+    "additionalOptions": "Lisavalikud",
+    "today": "Täna",
+    "yesterday": "Eile",
+    "last30days": "Viimased 30 päeva",
+    "months": "Valitud kuu",
+    "customperiod": "Valitud periood",
+    "until": "kuni"
+  },
   "menu": {
     "overview": "Ülevaade",
     "chats": "Vestlused",
